+++ src/datasources/dmphubAPI.ts
@@ -5,7 +5,7 @@ import {
 } from "@apollo/datasource-rest";
 import { logger, prepareObjectForLogs } from '../logger.js';
 import { DMPHubConfig } from '../config/dmpHubConfig.js';
-import { JWTAccessToken } from '../services/tokenService.js';
+import { JWTAccessToken } from "../types/general.js";
 import { MyContext } from "../context.js";
 import { RDACommonStandardContact, RDACommonStandardContributor } from "@dmptool/utils";
 import { isNullOrUndefined } from "../utils/helpers.js";